+++ CHANGES.md
@@ -13,6 +13,7 @@
 ### Deprecated
 ### Removed
 - Removed the Remix modules for Mists of Pandaria.
+- Removed the **Toggle All** button from the settings.
 ### Fixed
 - Resolved an issue that inhibited equipping identical heirloom items (e.g., two Swift Hands of Justice) with a single click of the **Equip Heirlooms** button.
 - Resolved an issue where heirlooms on max-level characters were being replaced by items with significantly lower item levels.